Pick Champ
Pick Champ is a fantasy sports website that is different than most
fantasy sports platforms.
The problem the client had was lack of automation to the competitions
and the client had no database,
 so the goal of this project is to automate the competitions as well as
create user analytics.
 The client wants to have an administrator able to sort through profile
characteristics such as gender, age and zip code.
The client also wanted the ability to sort and filter data by sports
team or user, as well as see the general user user statistics and track
what time of day people are entering the contests.